On branch master I have copied xmp_fh into winhomefs0.4.c I am using that and adding resolution  calls to it... this is bcause it's amore thurough file system implementation.... I will have to  remodify calls to read dir however....
 Changes to be committed:
	modified:   build.py
	modified:   fs.c
	modified:   fs.h
	modified:   hidden.c
	modified:   initialization.c
	modified:   utilities.c
	modified:   utilities.h
	modified:   winhomefs0.3.c
	new file:   winhomefs0.4.c

+++ hidden.c
@@ -33,7 +33,7 @@ list_t * get_hidden_list( const char * path ){
   
   if( hidden_f ){
     
-    int    size = 100;
+    size_t size = 100;
     char * buff = (char*)malloc(100);
     int    read = getline( &buff, &size, hidden_f );
